refactor: move sending signed transaction hook
