+++ src/app/content/Widget/StakeWidget.tsx
@@ -5,7 +5,7 @@ import { anticipate, motion } from "motion/react";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useTranslations } from "next-intl";
 import { useConnection, useWallet } from "@solana/wallet-adapter-react";
-import { isValidSolAmount, getMinimumStakeAmount } from "@/utils/solana";
+import { isValidSolAmount, getMinimumStakeAmount, useSendSignedTransaction } from "@/utils/solana";
 import { trpc } from "@/utils/trpc";
 import { Transaction } from "@solana/web3.js";
 import { Buffer } from "buffer";
@@ -35,9 +35,10 @@ export default function StakeWidget() {
     setTransactionStatus({ type: null, message: "", link: undefined });
   }, [setTransactionStatus]);
 
+  const sendSignedTransaction = useSendSignedTransaction();
+
   const t = useTranslations();
   const { publicKey, signTransaction, connected } = useWallet();
-  const { connection } = useConnection();
   const walletAddress = useMemo(
     () => publicKey?.toBase58() ?? "",
     [publicKey]
@@ -122,32 +123,6 @@ export default function StakeWidget() {
   const prepareDeactivateStakeTransaction = trpc.stake.prepareDeactivateStakeTransaction.useMutation();
   const prepareWithdrawStakeTransaction = trpc.stake.prepareWithdrawStakeTransaction.useMutation();
   const prepareMergeStakeTransaction = trpc.stake.prepareMergeStakeTransaction.useMutation();
-
-  const sendSignedTransaction = useCallback(async (signedTransaction: Transaction) => {
-    const serializedTransaction = signedTransaction.serialize();
-    const signature = await connection.sendRawTransaction(serializedTransaction, {
-      skipPreflight: false,
-      preflightCommitment: "confirmed",
-    });
-
-    const blockhashInfo = signedTransaction.recentBlockhash && signedTransaction.lastValidBlockHeight
-      ? {
-          blockhash: signedTransaction.recentBlockhash,
-          lastValidBlockHeight: signedTransaction.lastValidBlockHeight,
-        }
-      : await connection.getLatestBlockhash();
-
-    await connection.confirmTransaction(
-      {
-        signature,
-        blockhash: blockhashInfo.blockhash,
-        lastValidBlockHeight: blockhashInfo.lastValidBlockHeight,
-      },
-      "confirmed"
-    );
-
-    return signature;
-  }, [connection]);
 
   const tabs: Array<{ id: StakeTab; label: string }> = [
     { id: "stake", label: t("ui.stake") },
